requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.35.0
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==4.2.5
django-storages==1.13.1
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.5.0
psycopg2==2.9.7
django-heroku==0.3.1
gunicorn==21.2.0
Pillow==9.4.0
boto3==1.28.49
botocore==1.31.49
openai